% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ImportExport.R
\name{Download_renameExcel}
\alias{Download_renameExcel}
\title{Converts df to an Excel file}
\usage{
Download_renameExcel(ExcelFile, DownloadpathExcel)
}
\arguments{
\item{ExcelFile}{df, df that needs to be converted into an Excel file.}

\item{DownloadpathExcel}{string, File path}
}
\description{
Converts a df to an Excel file.
}
