% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ImportExport.R
\name{ExcelRead}
\alias{ExcelRead}
\title{Converts Excel file to a df}
\usage{
ExcelRead(Input_Read)
}
\arguments{
\item{Input_Read}{string, filepath to Excel file.}
}
\description{
Converts Excel into df.
}
