% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ImportExport.R
\name{ExportAllImages}
\alias{ExportAllImages}
\title{Save and export ggplot images based on df column}
\usage{
ExportAllImages(Columnparams, DownloadpathImages, loopname, loopid)
}
\arguments{
\item{Columnparams}{df column, column of df to base images on.}

\item{loopname}{loop, function is placed within a loop, references which loop to apply.}

\item{loopid}{value, letter to indicate iterations to loop through.}

\item{DownloadpatyImages}{string, file path chosen by the user.}
}
\value{
StandPlot, ggplot object used to export ggplot images.
}
\description{
Save and export plots based on df column, containing all possible combinations of variables.
}
