% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ImportExport.R
\name{Generate_renamedf}
\alias{Generate_renamedf}
\title{Create rename df based on column other df}
\usage{
Generate_renamedf(ColumntoRename)
}
\arguments{
\item{ColumntoRename}{column, column of df that is to be renamed.}
}
\value{
Old_Excel_Names, df contains two column: NewName & OldName. Both containing all unique values from original columns.
}
\description{
Creates a df based on the column that has to be renamed. The resulting df has an OldName and NewName column containing all unique values from the original column.
}
