% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ImportExport.R
\name{GenerateStandplotimage}
\alias{GenerateStandplotimage}
\title{Save ggplot object}
\usage{
GenerateStandplotimage(
  PlotvarName,
  Formatwidth,
  Formatheight,
  Formatunit,
  filecontent
)
}
\arguments{
\item{PlotvarName}{ggplot object, Name of ggplot.}

\item{Formatwidth}{numeric, Width of object (image).}

\item{Formatunit}{string, Unit of length to apply.}

\item{filecontent}{string, Name of file.}

\item{Formatheightnumeric, }{Height of object (image).}
}
\value{
StandPlot, ggplot object, used to export ggplot images.
}
\description{
Save ggplot object, used in application to generate ggplot images.
}
