% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Functions.R
\name{hline}
\alias{hline}
\title{Adds horizontal line to standard visualization plot}
\usage{
hline(Nameline, Lineselect, hlineyintcep, hlinecolor, hlinetype, hlinewitdh)
}
\arguments{
\item{Nameline}{string, name of selection box for hline.}

\item{Lineselect}{input of wigdet, input of selection box for hlines.}

\item{hlineyintcep}{numeric, Determines position of vline of y-axis.}

\item{hlinecolor}{string, input of text wigdet determines name of the hline.}

\item{hlinewidth}{numeric, width of the hline.}
}
\value{
hlineresult, hline to added to standard plot.
}
\description{
Adds horizontal line to standard visualization plot, using widgets as input.
}
