% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ImportExport.R
\name{ImportData}
\alias{ImportData}
\title{Get character vector of directory and file names}
\usage{
ImportData(FilePattern, DirSelect, NameFull, case_ignore, Recursive)
}
\arguments{
\item{FilePattern}{regular expression, Only listing files that contain the specified regular expression.}

\item{case_ignore}{logical, Should the function be case-sensitive.}

\item{Recursive}{logical, can the function search all directories.}

\item{DireSelect}{string, dirname() the output of the directory selector.}
}
\value{
listQC, directory and file names resulting from importing
}
\description{
Import data returns a character vector containing the names of the directories and files of the selected directory.
}
