% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Functions.R
\name{QCPlotFunctionDetail}
\alias{QCPlotFunctionDetail}
\title{Creates detailed visualization plot}
\usage{
QCPlotFunctionDetail(
  Finaldf,
  x_axis,
  y_axis,
  Instrument,
  Lasers_Detectors,
  widthline,
  limxmin,
  limxmax,
  limymin,
  limymax,
  titleplot,
  ylabaxis,
  xlabaxis,
  colorscale
)
}
\arguments{
\item{Finaldf}{df, df with contents that need to be plotted.}

\item{x_axis}{string, value from df presented on x-axis.}

\item{y_axis}{string, value from df presented on y-axis.}

\item{Instrument}{df column, uses the instrument column of df as input for linetype.}

\item{Lasers_Detectors}{string, output of the color filter controls what data to color in plot.}

\item{widthline}{number, Width of line in plot.}

\item{limxmin}{numeric, Minimumn of x-axis.}

\item{limxmax}{numeric, Maximum of x-axis.}

\item{limymin}{numeric, Minimum of y-axis.}

\item{limymax}{numeric, Minimum of y-axis.}

\item{titleplot}{string, Title of plot.}

\item{ylabaxis}{string, Title of y-axis.}

\item{xlabaxis}{string, Title of X-axis.}

\item{colorscale}{colorscale, Insert a color scale to use for plot.}
}
\value{
QCDetailPlot, ggplot object for detailed plot.
}
\description{
Generate ggplot for detailed visualization, does not create new scale for vlines.
}
