% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Functions.R
\name{QCPlotFunctionStand}
\alias{QCPlotFunctionStand}
\title{Creates standard visualization plot}
\usage{
QCPlotFunctionStand(
  Finaldf,
  x_axis,
  y_axis,
  Instrument,
  Lasers_Detectors,
  widthline,
  limxmin,
  limxmax,
  limymin,
  limymax,
  titleplot,
  ylabaxis,
  xlabaxis,
  colorscale,
  color1,
  color2,
  color3,
  color4,
  color5,
  color6,
  color7
)
}
\arguments{
\item{Finaldf}{df, df with contents that need to be plotted.}

\item{x_axis}{string, value from df presented on x-axis.}

\item{y_axis}{string, value from df presented on y-axis.}

\item{Instrument}{df column, uses the instrument column of df as input for linetype.}

\item{Lasers_Detectors}{string, output of the color filter controls what data to color in plot.}

\item{widthline}{number, Width of line in plot.}

\item{limxmin}{numeric, Minimumn of x-axis.}

\item{limxmax}{numeric, Maximum of x-axis.}

\item{limymin}{numeric, Minimum of y-axis.}

\item{limymax}{numeric, Minimum of y-axis.}

\item{titleplot}{string, Title of plot.}

\item{ylabaxis}{string, Title of y-axis.}

\item{xlabaxis}{string, Title of X-axis.}

\item{colorscale}{colorscale, Insert a color scale to use for plot.}

\item{color1}{string, color1 for vlines presented in the plot.}

\item{color2}{string, color2 for vlines presented in the plot.}

\item{color3}{string, color3 for vlines presented in the plot.}

\item{color4}{string, color4 for vlines presented in the plot.}

\item{color5}{string, color5 for vlines presented in the plot.}

\item{color6}{string, color6 for vlines presented in the plot.}

\item{color7}{string, color7 for vlines presented in the plot.}
}
\value{
QCplot, ggplot object for standard plot.
}
\description{
Generate ggplot for standard visualization, creates newscale for vline colors.
}
