% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Functions.R
\name{Varimageplot}
\alias{Varimageplot}
\title{Assign plot to global env}
\usage{
Varimageplot(ImageName, Varimage)
}
\arguments{
\item{ImageName}{string, Name object will take in global env.}

\item{varimage}{ggplot object, Name of ggplot object.}
}
\description{
Assigns a ggplot object to global env.
}
