% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Functions.R
\name{vline}
\alias{vline}
\title{Adds vertical line to standard visualization plot}
\usage{
vline(Nameline, Lineselect, vlinexintcep, vlinecolor, vlinetype, vlinewitdh)
}
\arguments{
\item{Nameline}{string, name of selection box for vline.}

\item{Lineselect}{input of wigdet, input of selection box for vlines.}

\item{vlinexintcep}{Date, Determines position of vline of x-axis.}

\item{vlinecolor}{string, input of text wigdet determines name of the vline.}

\item{vlinewidth}{numeric, width of vline.}
}
\value{
vlineresult, vline to added to standard plot.
}
\description{
Adds vertical line to standard visualization plot, using widgets as input.
}
