% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Functions.R
\name{vlinedetail}
\alias{vlinedetail}
\title{Adds vertical line to detailed visualization plot}
\usage{
vlinedetail(
  Nameline,
  Lineselect,
  vlinexintcep,
  vlinecolor,
  vlinetype,
  vlinewitdh,
  showlegd
)
}
\arguments{
\item{Nameline}{string, name of selection box for vline.}

\item{Lineselect}{input of wigdet, input of selection box for vlines.}

\item{vlinexintcep}{Date, Determines position of vline of x-axis.}

\item{vlinecolor}{string, input of text wigdet determines name of the vline.}

\item{vlinewidth}{numeric, width of vline.}

\item{showlegend}{logical, show vline in legend, recommened set to TRUE.}
}
\value{
vlineresultdetail, vline to added to detailed plot.
}
\description{
Adds vertical line to detailed visualization plot, using widgets as input.
}
